feat: update request and response structures to use DTOs in authentication handlers and services
